Adding some runner scripts

+++ include/config.h
@@ -7,7 +7,7 @@
 #if USE_STRING_KEYS
 
 typedef Slice KEY_TYPE;
-typedef uint64_t PLR_SEGMENT_POINT;
+typedef double PLR_SEGMENT_POINT;
 
 #elif USE_INT_128
